+++ src/Entity/Traits/UpdatedAtTrait.php
@@ -1,23 +1,12 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Entity\Traits;
 
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
-trait TimestampableTrait
+trait UpdatedAtTrait
 {
-    #[ORM\Column(
-        name: 'created_at',
-        type: 'datetime_immutable',
-        nullable: true,
-        updatable: false
-    )]
-    #[Gedmo\Timestampable(on: 'create')]
-    private ?\DateTimeImmutable $createdAt = null;
-
     #[ORM\Column(
         name: 'updated_at',
         type: 'datetime_immutable',
@@ -26,15 +15,6 @@ trait TimestampableTrait
     #[Gedmo\Timestampable(on: 'update')]
     private ?\DateTimeImmutable $updatedAt = null;
 
-    public function getCreatedAt(): ?\DateTimeImmutable
-    {
-        return $this->createdAt;
-    }
-
-    public function setCreatedAt(\DateTimeImmutable $createdAt): void
-    {
-        $this->createdAt = $createdAt;
-    }
 
     public function getUpdatedAt(): ?\DateTimeImmutable
     {
